telegram-kilo-code-automation/requirements.txt
# Dépendances pour l'automatisation Telegram -> Kilo Code VSCode
python-telegram-bot==20.7
pyautogui==0.9.54
python-dotenv==1.0.0
pynput==1.7.6
pillow==10.1.0
